requirements.txt
cffi==1.12.3
cryptography==3.4.7
pycparser==2.19
pyOpenSSL==20.0.1
six==1.15.0
